requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.24.46
botocore==1.27.46
cached-property==1.5.2
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==3.2.14
django-cacheops==6.1
django-celery-results==2.4.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-redis==5.2.0
django-seed==0.3.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
Faker==12.1.0
funcy==1.17
idna==3.3
importlib-metadata==4.10.1
jmespath==1.0.1
kombu==5.2.4
Markdown==3.3.6
mysqlclient @ file://mysqlclient-1.4.6-cp37-cp37m-win_amd64.whl
packaging==21.3
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.3
redis==4.3.4
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toposort==1.7
typing_extensions==4.0.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.7.0
